+++ src/Readable.php
@@ -2,27 +2,27 @@
 /**
  * Trait with utility methods to handle errors.
  *
- * PHP version 5.4
+ * PHP version 5.6
  *
  * @author Francesco Bianco
  */
-namespace Javanile\SchemaDB;
+namespace Javanile\Moldable;
 
-use Javanile\SchemaDB\Notations;
+use Javanile\Moldable\Notations;
 
 class Readable implements Notations
 {
-    use Model\LoadApi;
-    use Model\ReadApi;
-    use Model\JoinApi;
-    use Model\TableApi;
-    use Model\FieldApi;
-    use Model\FetchApi;
-    use Model\ClassApi;
-    use Model\SchemaApi;
     use Model\DatabaseApi;
-    use Model\PublicApi;
-    use Model\DebugApi;
+    #use Model\LoadApi;
+    #use Model\ReadApi;
+    #use Model\JoinApi;
+    #use Model\TableApi;
+    #use Model\FieldApi;
+    #use Model\FetchApi;
+    #use Model\ClassApi;
+    #use Model\SchemaApi;
+    #use Model\PublicApi;
+    #use Model\DebugApi;
     //
     //use Model\ErrorApi;
     //
